Moved solveQp to HighsSolve
